move get_color_bar_and_data_ranges to plotting._utils
