+++ nilearn/plotting/surf_plotting.py
@@ -24,13 +24,13 @@ from nilearn.plotting._utils import (
     check_surface_plotting_inputs,
     create_colormap_from_lut,
     get_cbar_ticks,
+    get_colorbar_and_data_ranges,
     sanitize_hemi_for_surface_image,
     save_figure_if_needed,
 )
 from nilearn.plotting.cm import mix_colormaps
 from nilearn.plotting.displays import PlotlySurfaceFigure
 from nilearn.plotting.html_surface import get_vertexcolor
-from nilearn.plotting.img_plotting import get_colorbar_and_data_ranges
 from nilearn.plotting.js_plotting_utils import colorscale
 from nilearn.surface import (
     load_surf_data,
