+++ nilearn/plotting/tests/test_utils.py
@@ -5,7 +5,7 @@
 import numpy as np
 import pytest
 
-from nilearn.plotting.img_plotting import get_colorbar_and_data_ranges
+from nilearn.plotting._utils import get_colorbar_and_data_ranges
 
 
 @pytest.fixture
